+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+## Release (XXXX-XX-XX)
+
+- `lbapplication`: [v0.2.0](services/lbapplication/CHANGELOG.md#v020-2025-01-30)
+  - **Bugfix**: Set type from interface to int64 of `HealthyThreshold`, `UnhealthyThreshold`, `MaxLoadBalancers`, `Port`, `MaxConnections`, `Code` and `TargetPort`
+
+
 ## Release (2025-01-24)
 
 - `ske`: [v0.21.1](services/ske/CHANGELOG.md#v0211-2025-01-23)