+++ services/lbapplication/CHANGELOG.md
@@ -1,3 +1,7 @@
+## v0.2.0 (2025-01-30)
+
+- **Bugfix**: Set type from interface to int64 of `HealthyThreshold`, `UnhealthyThreshold`, `MaxLoadBalancers`, `Port`, `MaxConnections`, `Code` and `TargetPort`
+
 ## v0.1.0 (2025-01-13)
 
 - **New**: STACKIT Load Balancer Application module can be used to manage the STACKIT Load Balancer Application